Filter combine

Filters a data frame then binds the result with it (with names).

Usage

filter_combine(x, ..., .id, .names)

Arguments

x

A data.frame

...

Arguments passed on to dplyr::filter

.by

<tidy-select> Optionally, a selection of columns to group by for just this operation, functioning as an alternative to group_by(). For details and examples, see ?dplyr_by.

.id

When .id is supplied, a new column of identifiers is created to link each row to its original data frame. The labels are taken from `.names`

.names

vector of characters for the

See also